requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
h11==0.16.0
idna==3.11
numpy==2.4.1
pandas==2.2.3
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
chromadb==1.1.0
uvicorn==0.40.0
